[backport] Prevent failed lease update from setting lease status to ERROR
If a lease update fails because of not enough hosts being available
(either when prolonging or deferring a lease, increasing the number of
reserved hosts, or changing the requirements), it would leave the lease
in ERROR status, although the reservation would still be active.

This patch allows the lease status code to restore the original lease
status after the failed update.

Change-Id: Id286e5a0f9f278cbf5330887bfea1719cd25ae1a
Closes-Bug: #1786031