[backport] Stop using trusts to get context for notifications
Blazar retrieves a context object from the trust recorded when a lease
was created and sends it as part of notification messages. However, if
the user who created the lease is deactivated or is not a member of the
project anymore, operations modifying the lease and lease events fail to
execute with an AuthorizationFailure exception.

This is because the trust object is not valid anymore due to the change
of permissions for the trustor.

This patch removes the context from all notification messages and stops
using trusts except for the initial lease creation. This allows project
members and OpenStack admins to manage the lease after the creator user
is disabled or has been removed from the project, as well as lease
events to run properly.

Change-Id: I895e9a54948ff5fd4dbfdad71c64631d71ef98fb
Closes-Bug: #1712381