Only return 1 allocation candidate for network res
Allocation candidates are supposed to be candidates which will
be directly reserved as part of a lease, not all possible candidates.
Filter the _matching_networks call down to a list of length 1, which
is the current (non-configurable) amount of networks which can be
reserved as part of a particular network reservation.

This also updates it to raise an exception if the number of
networks found was <1

Change-Id: Ia2bf660e6f5a316975ef7b9d2c33a61d71b65f78