keep enforcement filters in order
Currently, the enforcement filters don't run in order. So it may cause
allocation overcharge when the balance service filter runs before the
max_reservation_length filter. For example, if someone tries to extend
the lease two days prior to the end time, the project will be charged
with the encumbered SUs but then failed on extending the lease.