Handle k8s node rebuild case
If a K8s node for whatever reason is rebuilt, it will no longer
have the "device" label that Blazar uses to find it. This means
it will never mark the node as healthy again.

To resolve this, we can instead always look up the node by name. The
name is not mutable so it should always correspond. If the node is
rebuilt and thus missing the label, we can also re-add the label
when this is detected.

Change-Id: I6ccff49e4e170287a4af55418ed2d27c8ef9f31d